Slightly better workaround for truncation issue
Even if you add one or two spaces, the '.' character is still truncated. Also, if you have a function with no arguments, then it is the function help that is truncated. I've found that adding an empty string argument as the last argument of a given function registration is a better workaround. Nothing is truncated, and the '.' character is shown in all cases.