Update pidtagservicedllname-canonical-property.md
Via CI 100773, this request is from Outlook PG to delete the sentence "The base filename contains up to six characters that uniquely identify the DLL." because MAPI does not limit the filename to 6 characters and having this incorrect limit in place may cause issues for some users especially now that MAPI is being more restrictive about the DLLs that it will load. Would you please approve this update? Thanks!