Update lookuprecord-data-block.md
I noticed that the **LookupRecord** documentation lacked any description of what occurs when no records match the specified criteria.  I found an answer on Stack Exchange (https://stackoverflow.com/questions/29013731/data-macros-using-default-values-if-lookuprecord-does-not-find-a-match/29029441#29029441), personally validated it in Access 2016, proposed it here so that it can become part of the official, publicly available Microsoft documentation.  Thank you for your time and consideration.