Update compatibility-between-the-32-bit-and-64-bit-versions-of-office.md
It's very confusing to have the statement "By default, when you install a 64-bit version of Office you also install the 32-bit version alongside it." as a Note at the top of the page and then "Can I install 64-bit and 32-bit versions of Office side-by-side? No." at the bottom.

If you are referencing something else than Office (like VBA 7 that is mentioned above and below the Note) then this has to be explicitly called out.