MAPI: mention PT_FILE_HANDLE type name
https://github.com/stephenegriffin/mfcmapi/blob/master/include/EdkMdb.h#L1228
mentions the PT_FILE_HANDLE type, also of value 259:

>PT_FILE_HANDLE: real data is in file specified by handle.
>prop.Value.l has file handle

Value.l is a fixed 32-bit entity and insufficient to hold a
[[HANDLE]](https://docs.microsoft.com/en-us/windows/win32/winprog/windows-data-types)
in the LP64 data model. However, it can assumed that the described use of
Value.l is owed to a time predating 64-bit Windows.

Value.lpv is the one that makes sense for ILP32 and LP64.
Therefore, add just the PT alias.